connectivity change handling looks very solid in iOS.  Not sure what I added will help at all, looks like you are/were doing it the recommended way.  Got rid of a bunch of string format/logging warnings

+++ surespot/libs/SDWebImage/SDWebImageDownloaderOperation.mm
@@ -249,7 +249,7 @@ static const int ddLogLevel = LOG_LEVEL_OFF;
 
 - (void)connection:(NSURLConnection *)connection didReceiveData:(NSData *)data
 {
-    DDLogInfo(@"appending %d bytes", data.length);
+    DDLogInfo(@"appending %lu bytes", (unsigned long)data.length);
     [self.imageData appendData:data];
     
     //    if ((self.options & SDWebImageDownloaderProgressiveDownload) && self.expectedSize > 0 && self.completedBlock)
